Add block operator functions.
The new block module contains functions to achieve block operators. The
bmat function converts a list of lists [[LinearOperator]] into a single
new block operator.

bmat assumes the input is a list of rows of a block matrix.

Various internal functions are used to the construct the new block
operator and are called by bmat. Functions named hstack and vstack are
used in this context of LinearOperators and are similar to numpy hstack
and vstack.