Add ensure2dColumn decorator.
The ensure2dColumn decorator takes a function defined in the following
format

  def func(op_shape, x)

and ensures that x is at least 2D, converting 1D arrays to 2D column
arrays if required. Additionally, 2D results from the computation of
func that are only a single column are converted into 1D arrays.
This respects func's input, as defined by the Numpy code guidelines.