Fix py27 test fail for ensure2dColumn.
The error occurs when the wrapped function is the result of partial.
This is due to a bug in Python 2.7 that was fixed in Python 3.3. More
information can be found here

  http://bugs.python.org/issue3445

A side effect of this is that, in Python 2, the functions returned from
operators that used this `ensure2dColumn(partial(...))` pattern
(currently zeros and ones) do not have their associated docstring/etc,
since the functools.wraps decorator is not used.