LinearOperator __call__ now checks all shapes.
The shape of the input and the result are now interegated in order to
make sure that they match the dimensionality required for the package.
The input is not allowed to be zero dimensional, and is not allowed to
be a tensor of order greater than 2. The same rules apply to the result
of the forward function.

Additionally, if the expected shape of the forward application
(self._shape[0], x.shape[1]) is not held, then LinearOperator throws a
DimensionMismatch.