Fix ensure2dColumn: matrix to vector singleton.
Previously ensure2dColumn tried to check for singletons based on the
output, but there was a case in the nullspace of this check. By checking
based on the input, the problem is resolved.

In particular for res = [[0]] the return was always [[0]] previously.
However, we actually want to return [[0]] only if the input was a matrix
(2D array) and to return [0] if the input was a vector (1D array).