Fix motivation.rst not compiling under latexpdf.
This was due to the use of the align environment. The .. math:: directive
automatically adds a split environment around the code given in the
directive. At the end, the problem was \begin{split}\begin{align*} ...
\end{align*}\end{split}\notag cannot distinguish between split and align
environment (this is a guess), and fails to figure out where the notag
is a reference to.